# Conflicts:
#	frontend/src/components/Header.css
#	frontend/src/components/WeatherWidget.css
#	frontend/src/components/WeatherWidget.jsx

+++ backend/src/main/java/fer/progi/backend/service/impl/NastavnikServiceJpa.java
@@ -2,8 +2,6 @@ package fer.progi.backend.service.impl;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -66,17 +64,23 @@ public class NastavnikServiceJpa implements NastavnikService{
 	}
 
 	@Override
-	public void deleteNastavnik(String email) {
+	public boolean deleteNastavnik(String email) {
 		Nastavnik nastavnik = nastavnikRepo.findByEmail(email).orElse(null);
-		nastavnikRepo.delete(nastavnik);
+		if (nastavnik != null && (nastavnik.getPredmeti() == null || nastavnik.getPredmeti().isEmpty())) {
+			nastavnikRepo.delete(nastavnik);
+			return true;
+		}
+		return false;
 	}
-	
+
+
 	@Override
-	public Set<Predmet> findNastavnikPredmeti(String email) {
+	public List<Predmet> findNastavnikPredmeti(String email) {
 		Nastavnik nastavnik = nastavnikRepo.findByEmail(email)
 	            .orElseThrow(() -> new RuntimeException("Nastavnik nije pronađen s emailom: " + email));
-	        
-	        return nastavnik.getPredmeti();
+
+		System.out.println("Predmeti: " + nastavnik.getPredmeti());
+		return nastavnik.getPredmeti();
 	}
 		
 }
